requirements.txt
asgiref==3.7.2
contourpy==1.3.3
cycler==0.12.1
Django==4.2.10
django-bootstrap5==23.4
djangorestframework==3.16.1
dnspython==2.8.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.9
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
seaborn==0.13.2
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2025.2